feat: allow query to be executed inside a transaction
