--- camus/__init__.py
@@ -1,3 +1,0 @@
-from .database import Database
-
-__version__ = "0.0.2"
